Add some Excel import/export features

+++ app/Http/Livewire/Attendance/Gathering/Edit.php
@@ -3,27 +3,36 @@
 namespace App\Http\Livewire\Attendance\Gathering;
 
 use Livewire\Component;
+use Livewire\WithFileUploads;
+use App\Imports\GatheringImport;
+use Maatwebsite\Excel\Facades\Excel;
+use Illuminate\Support\Facades\Storage;
 use App\Models\Gathering;
 use App\Models\User;
 use Carbon\Carbon;
 
 class Edit extends Component
 {
+	use WithFileUploads;
+	
 	public $gath, $name, $held_at, $description;
-	public $teachers;
+	public $teachers, $excel, $manual_input = 'manual';
 	public $ids = [];
 	
-	protected $listeners = ['sendTeacher' => 'setTeachers'];
+	protected $listeners = ['sendTeacher' => 'setTeachers', 'reRender' => '$refresh'];
 	
 	protected $rules = [
 		'held_at'	=> 'required|date_format:d/m/Y',
 		'name'	=> 'required',
+		'excel'	=> 'nullable|file|mimes:xls,xlsx'
 	];
 	
 	protected $messages = [
 		'held_at.required'	=> 'Tanggal kegiatan tidak boleh kosong.',
 		'held_at.date_format'	=> 'Format penulisan tanggal salah.',
 		'name.required'	=> 'Nama kegiatan tidak boleh kosong.',
+		'excel.file'	=> 'Format file tidak valid.',
+		'excel.mimes'	=> 'Hanya file Excel yang dapat diupload.'
 	];
 	
 	public function setTeachers($teachers){
@@ -55,20 +64,40 @@ class Edit extends Component
 			'held_at'	=> Carbon::createFromFormat('d/m/Y', $this->held_at)->format('Y-m-d'),
 			'description'	=> $this->description,
 		]);
-		if(!empty($this->teachers)) {
-			$ids = array();
-			foreach ($this->teachers as $t) {
-				$ids[] = $t['id'];
+		if($this->manual_input == 'manual'){
+			if(!empty($this->teachers)) {
+				$ids = array();
+				foreach ($this->teachers as $t) {
+					$ids[] = $t['id'];
+				}
+				$this->gath->users()->sync($ids);
+			} else {
+				$this->gath->users()->detach();
 			}
-			$this->gath->users()->sync($ids);
 		} else {
-			$this->gath->users()->detach();
+			if($this->excel){ 
+				$ids = $this->importExcel();
+				$this->gath->users()->sync($ids);
+			}
 		}
 		$this->emit('saved');
+		return redirect()->route('gathering.edit', $this->gath->id);
+		
 	}
 	
 	public function render()
 	{
 		return view('livewire.attendance.gathering.edit');
+	}
+	
+	public function importExcel(){
+		$import =  new GatheringImport($this->gath);
+		Excel::import($import, $this->excel);
+		$this->reset('excel');
+		return $import->getIds();
+	}
+	
+	public function downloadTemplate(){
+		return Storage::disk('public')->download('excel/TEMPLATE_KUMPUL.xlsx', 'TEMPLATE_KUMPUL_' . time() . '.xlsx');
 	}
 }
